requirements.txt
aiodns==3.3.0
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
defusedxml==0.7.1
Django==5.2
django-environ==0.12.0
django-jazzmin==3.0.1
idna==3.10
oauthlib==3.2.2
pillow==11.2.1
pycares==4.8.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
sslcommerz-lib==1.0
sslcommerz_python_api==1.0.1
tzdata==2025.2
urllib3==2.4.0
